Using CSS Modules
In this commit the CSS modules was implemented in the Button.cs so the file name changed to Button.module.css and it was imported using - import styles from ./Button.module.css; The Css Modules is able to change the css class names to unique names such Button_button__KXFVC avoiding the overlap of css class names. The className was set to styles.button in the button.js. This is the best choice because the css code stay separately and the classNames generated to link with css by the CSS modules are uniques